Streamline admin settings stack
